requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
pdf2image==1.17.0
Pillow==10.4.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pdfplumber==0.11.4